task5_7
task5_7

+++ client.js
@@ -3,6 +3,8 @@ const net = require('net');
 const port = 8124;
 const client = new net.Socket();
 const requestString = '\r\nQA\r\n';
+const ascString = '\r\nASC\r\n';
+const decString = '\r\nDEC\r\n';
 
 Array.prototype.shuffle = function( b )
 {
@@ -23,18 +25,37 @@ function random(){
     return json.items.shuffle();
 }
 
+let items = random();
+
 client.setEncoding('utf8');
 
 client.connect(port, function() {
     console.log('Connected');
-    let items = random();
-    console.log(items);
     client.write(requestString);
 });
 
 client.on('data', function(data) {
-    console.log(data);
-    client.destroy();
+    if (items.length > 0){
+        if (data === ascString){
+            client.id = items.pop();
+            client.write(client.id["request"]);
+        }else if (data === decString){
+            client.destroy();
+        }else{
+            let response = "";
+            if (data === client.id["response"]){
+                response = "OK";
+            }else{
+                response = "ERROR";
+            }
+            console.log(client.id);
+            console.log(response);
+            client.write(requestString);
+        }
+    }else{
+        client.destroy();
+    }
+
 });
 
 client.on('close', function() {
